+++ application/resources/js/Components/SearchBar.tsx
@@ -1,5 +1,5 @@
 import TextInput from '@/Components/atoms/TextInput';
-import useEscapeKey from '@/Hooks/useEscapeKey';
+import useEscapeKey from '@/useHooks/useEscapeKey';
 import { router } from '@inertiajs/react';
 import { useLaravelReactI18n } from 'laravel-react-i18n';
 import React, {