+++ application/resources/js/Pages/Charts/Partials/LineChart.tsx
@@ -4,7 +4,7 @@ import ArrowTrendingDown from '@/Components/svgs/ArrowTrendingDown';
 import ArrowTrendingUp from '@/Components/svgs/ArrowTrendingUp';
 import { useFilterContext } from '@/Context/FiltersContext';
 import { userSettingEnums } from '@/enums/user-setting-enums';
-import { useUserSetting } from '@/Hooks/useUserSettings';
+import { useUserSetting } from '@/useHooks/useUserSettings';
 import { shortNumber } from '@/utils/shortNumber';
 import { ResponsiveLine } from '@nivo/line';
 import { useLaravelReactI18n } from 'laravel-react-i18n';